% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backpropagate.R
\name{backpropagate}
\alias{backpropagate}
\title{Calculate the Error and Backpropagate adjusts}
\usage{
backpropagate(input, output, network)
}
\arguments{
\item{input}{A matrix of size (N, 1), where N is the size
of the layer that is going to receive such inputs. (matrix)}

\item{output}{A matrix of size (M, 1) representing the
expected outputs given some inputs, where M is the network
output size. (matrix)}

\item{network}{A network object generated by create_network (list)}
}
\description{
This function does a backwards propagation.
}
\author{
Eduardo Kapp
}
