% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_layer.R
\name{compute_layer}
\alias{compute_layer}
\title{Computer layer outputs}
\usage{
compute_layer(inputs, layer)
}
\arguments{
\item{inputs}{A matrix of size (N, 1), where N is the size
of the layer that is going to receive such inputs. (matrix)}

\item{layer}{A layer object generated by create_layer (list)}
}
\description{
This function returns an array of outputs
from a given layer.
}
\author{
Eduardo Kapp
}
