% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_layer.R
\name{create_layer}
\alias{create_layer}
\title{Create a network layer}
\usage{
create_layer(size, n_inputs, weights, biases, activation = sigmoid)
}
\arguments{
\item{size}{The layer size, which is the number of neurons (integer)}

\item{n_inputs}{The number of inputs to each neuron in the layer (integer)}

\item{weights}{A weight array associated with each neuron. Note that
the size of this array should be (size, n_inputs), where size is the
number of neurons and n_inputs is the number of inputs to this
layer. (matrix)}

\item{biases}{A bias array associated with each neuron. Note that
the size of this array should be (size, 1), where size is the number
of neurons. (matrix)}

\item{activation}{What activation function should be used. Default: sigmoid
(function)}
}
\value{
A network layer object represented as a list.
}
\description{
This function generates a layer object
represented as a list. The list contains all the
information about the layer, such as number of inputs,
weights and biases for every neuron in it.
}
\author{
Eduardo Kapp
}
