% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{create_network}
\alias{create_network}
\title{Create a neural network base structure}
\usage{
create_network(
  input_size,
  n_layers = 1,
  layer_sizes = c(1),
  activations = c("sigmoid")
)
}
\arguments{
\item{input_size}{The network input size (integer)}

\item{n_layers}{Number of hidden layers to be used. Note that this
includes the output layer. (integer)}

\item{layer_sizes}{An array with the number of neurons for each layer.
For example, if there are two layers, each with 30 neurons,
then layer_sizes must be c(30, 30). (integer array)}

\item{activations}{An array indicating what activation functions should be
used for each layer. By default, sigmoid functions are used. (character)}
}
\value{
A network layer object represented as a list.
}
\description{
This function generates a network object
represented as a list. The list contains all the
information about the network, such as input size, output size,
number of hidden layers, hidden layers sizes.
}
\author{
Eduardo Kapp
}
