% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_pass.R
\name{forward_pass}
\alias{forward_pass}
\title{Forward pass a set of inputs throughout the network}
\usage{
forward_pass(inputs, network)
}
\arguments{
\item{inputs}{A matrix of size (N, 1), where N is the size
of the layer that is going to receive such inputs. (matrix)}

\item{network}{A network object generated by create_network (list)}
}
\description{
This function does a forward propagation.
}
\author{
Eduardo Kapp
}
