% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(y, h)
}
\arguments{
\item{y}{An array (matrix)}

\item{h}{An array (matrix)}
}
\description{
An alternative MSE implementation where
the mean is halved. That is done in order to make the
derivative easier to calculate in the backpropagation
algorithm, as the derivative of 1/2n*(sum(y-h)^2) is then
simply (h-y).
}
\details{
Note that y and h need to have the same
dimensions.
}
\author{
Eduardo Kapp
}
