% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse_prime}
\alias{mse_prime}
\title{Derivative of the Mean Squared Error}
\usage{
mse_prime(y, h)
}
\arguments{
\item{y}{An array (matrix)}

\item{h}{An array (matrix)}
}
\description{
This implementation is the derivative of the alternate MSE.
}
\details{
Note that y and h need to have the same
dimensions.
}
\author{
Eduardo Kapp
}
