% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dnn}
\alias{predict.dnn}
\title{A neural network prediction function}
\usage{
\method{predict}{dnn}(model, data = X.test)
}
\arguments{
\item{model}{takes a list of learned parameters D=Nfeatures, K=Ncatagories, H=Nhidden layers, W1, b1, W2, b2}

\item{data}{A data set}
}
\description{
This function allows you to make predictions based on a trained NN model.
}
\examples{
predict.dnn(model, testdata)

example with DNN package

samp <- c(sample(1:50, 25), sample(51:100, 25), sample(101:150, 25))

iris.model <- train.dnn(x=1:4,y=5, traindata=iris[samp,],
                     testdata=iris[-samp,],hidden=10,
                     maxit=2000,display=50)

labels.dnn <- predict.dnn(iris.model, iris[-samp, -5])

# Display a confusion matrix of the dataset
table(iris[-samp, 5], labels.dnn)
}
\keyword{DNN,}
\keyword{NN}
\keyword{deep}
\keyword{network,}
\keyword{neural}
