% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relu.R
\name{relu}
\alias{relu}
\title{Rectified Linear Unit}
\usage{
relu(inputs)
}
\arguments{
\item{inputs}{An array}
}
\value{
An array with the same size of the input array.
}
\description{
Activation functin based on a half-rectifier.
}
\author{
Eduardo Kapp
}
