% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Activation Unit}
\usage{
sigmoid(inputs)
}
\arguments{
\item{inputs}{An array.}
}
\value{
An array with the same size of the inputs.
}
\description{
Activation functin based on a half-rectifier.
}
\author{
Eduardo Kapp
}
