% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid_prime}
\alias{sigmoid_prime}
\title{Sigmoid Activation Unit Derivative}
\usage{
sigmoid_prime(inputs)
}
\arguments{
\item{inputs}{An array.}
}
\value{
An array with the same size of the inputs.
}
\description{
This is a recursive implementation of the sigmoid derivative.
}
\author{
Eduardo Kapp
}
