% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.dnn}
\alias{train.dnn}
\title{A neural network training function}
\usage{
train.dnn(x, y, traindata = data, testdata = NULL, model = NULL,
  hidden = c(6), maxit = 2000, abstol = 0.01, lr = 0.01, reg = 0.001,
  display = 100, random.seed = 1)
}
\arguments{
\item{x}{columns from data to use as input}

\item{y}{column from data to use for classification}

\item{traindata}{data to train on}

\item{testdata}{data to test on Defaults to NULL}

\item{model}{a supplied model Defaults to NULL}

\item{hidden}{number of layers Defaults 6}

\item{maxit}{max iterations Defaults to 2000}

\item{abstol}{delta parameter for the loss/cost function Defaults to .01}

\item{lr}{learning rate Defaults to .01}

\item{reg}{rate of regularization Defaults to .001}

\item{display}{progress of learned parameters by steps Defaults to 100}

\item{random.seed}{set a seed value Defaults to 1}
}
\description{
This function allows you to train a NN.
}
\examples{
train.dnn(x=1:2, y=1,trainingdata=d.train,testdata=d.test,hidden=6,maxit=2000,display=100)

example with DNN package
samp <- c(sample(1:50, 25), sample(51:100, 25), sample(101:150, 25))
iris.model <- train.dnn(x=1:4,y=5, traindata=iris[samp,],
                     testdata=iris[-samp,],hidden=10,
                     maxit=2000,display=50)

labels.dnn <- predict.dnn(iris.model, iris[-samp, -5])

# Display a confusion matrix of the dataset
table(iris[-samp, 5], labels.dnn)
}
\keyword{DNN,}
\keyword{NN}
\keyword{deep}
\keyword{network,}
\keyword{neural}
