% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_network.R
\name{train_network}
\alias{train_network}
\title{Train a Network using Stochastic Gradient Descent}
\usage{
train_network(
  network,
  train_x,
  train_y,
  batch_size,
  iters = 1000,
  eta = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{network}{A kappnet network object. (list)}

\item{train_x}{A matrix with inputs for the training examples. Its
dimensions are expected to be (input_size, number_of_observations)(matrix).}

\item{train_y}{A matrix with outputs for the training examples. Its
dimensions are expected to be (output_size, number_of_observations)(matrix).}

\item{batch_size}{The batch size. Note that batch size does not mean input
size, but number of examples or observations that are used to calculate
a round of weights and biases adjustment. (integer)}

\item{iters}{Number of iterations. Default: 1000 (integer)}

\item{eta}{The learning rate for the gradient descent.
Default: 0.1 (numeric)}

\item{verbose}{Whether or not the algorithm should print out
information during its run time. Default: TRUE (logical)}
}
\value{
A trained network.
}
\description{
This function updates the network's weights and biases
by using the Stochastic Gradient Descent, a faster approach to the regular
Gradient Descent. The idea is that, for every iteration, training data is
shuffled and a number of batches are generated according to batch_size.
Then, the cost function is evaluated for every batch, thus updating the
weights and biases. This process is repeated for the number of iterations.
}
\details{
Note that train_x and train_y must be "aligned", that is, the first
column of train_x should relate to the first column of train_y, given that
the columns represent observations.
}
\author{
Eduardo Kapp
}
